pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "manim-banim"
version = "0.1.0"
description = "Molecular biology animation primitives for Manim"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "manim>=0.18.0",
]

[project.optional-dependencies]
pipeline = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pymupdf>=1.24.0",
    "azure-cognitiveservices-speech>=1.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "arxiv>=2.1.0",
    "python-dotenv>=1.0.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.entry-points."manim.plugins"]
"manim_banim" = "manim_banim"

[tool.setuptools.packages.find]
where = ["src"]
